+++ packages/eslint-plugin/typings/eslint.d.ts
@@ -1,7 +1,8 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
 // module augmentation is weird
 import { Scope } from 'eslint';
 declare module 'eslint' {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   namespace Scope {
     interface Variable {
       eslintUsed: boolean;